Dice id as long

+++ backend/src/main/kotlin/dinf/backend/DBDice.kt
@@ -5,7 +5,7 @@ import dinf.domain.SerialNumber
 import dinf.exposed.DiceEntity
 
 class DBDice(entity: DiceEntity) : Dice by Dice.Simple(
-    serialNumber = SerialNumber.Simple(entity.id.value.toLong()),
+    serialNumber = SerialNumber.Simple(entity.id.value),
     name = DBDiceName(entity),
     edges = DBEdges(entity)
 )
